requirements.txt
anaconda-anon-usage @ file:///C:/b/abs_c3w_h1zzjg/croot/anaconda-anon-usage_1710965204622/work
archspec @ file:///croot/archspec_1709217642129/work
asgiref==3.8.1
bokeh==3.6.2
boltons @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/boltons_1699480450092/work
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
certifi @ file:///C:/b/abs_7bbxxprxir/croot/certifi_1720453609121/work/certifi
cffi @ file:///C:/b/abs_78eb1_vq6z/croot/cffi_1714483206096/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1699472650914/work
conda @ file:///C:/b/abs_69e_za3rd5/croot/conda_1722004627938/work
conda-content-trust @ file:///C:/b/abs_bdfatn_wzf/croot/conda-content-trust_1714483201909/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1721662679737/work/src
conda-package-handling @ file:///C:/b/abs_831_pypuuz/croot/conda-package-handling_1718138287790/work
conda_package_streaming @ file:///C:/b/abs_fe6zb0_j8_/croot/conda-package-streaming_1718136098642/work
contourpy==1.3.1
cryptography @ file:///C:/b/abs_97lgskwerx/croot/cryptography_1714664736250/work
distro @ file:///C:/b/abs_71xr36ua5r/croot/distro_1714488282676/work
Django==5.1.2
django-filter==24.3
django-rest-framework==0.1.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozendict @ file:///C:/b/abs_2alamqss6p/croot/frozendict_1713194885124/work
gunicorn==23.0.0
h11==0.14.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
Jinja2==3.1.4
jsonpatch @ file:///C:/b/abs_4fdm88t7zi/croot/jsonpatch_1714483974578/work
jsonpointer==2.1
libmambapy @ file:///C:/b/abs_89suv9zwkz/croot/mamba-split_1714484512428/work/libmambapy
Markdown==3.7
MarkupSafe==3.0.2
menuinst @ file:///C:/b/abs_76jwd47ck1/croot/menuinst_1723567618942/work
mysqlclient==2.2.5
numpy==2.1.3
packaging @ file:///C:/b/abs_c3vlh0z4jw/croot/packaging_1720101866539/work
pandas==2.2.3
pillow==11.0.0
platformdirs @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/platformdirs_1701797392447/work
plotly==5.24.1
pluggy @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pluggy_1699472504117/work
pycosat @ file:///C:/b/abs_5csdern___/croot/pycosat_1714513102923/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT==2.9.0
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests @ file:///C:/b/abs_9frifg92q2/croot/requests_1721410901096/work
ruamel.yaml @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/ruamel.yaml_1699483184324/work
setuptools==72.1.0
six==1.17.0
sqlparse==0.5.1
tenacity==9.0.0
tornado==6.4.2
tqdm @ file:///C:/b/abs_a7h7n45ccq/croot/tqdm_1716395987133/work
truststore @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/truststore_1701881385424/work
typing_extensions==4.12.2
tzdata==2024.2
urllib3 @ file:///C:/b/abs_a7hvzm4y95/croot/urllib3_1718912661242/work
uvicorn==0.32.1
wheel==0.43.0
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
xyzservices==2024.9.0
zstandard @ file:///C:/b/abs_e46tf0_91g/croot/zstandard_1714678303615/work